+++ libs/skill-pdf2pic/rollup.config.cjs
@@ -21,8 +21,8 @@ module.exports = withNxDigipair(['esm', 'cjs'], config =>
     {      
       external: (id) => {
         const deps = config.format === 'cjs'
-          ? ["pdf2pic"]
-          : ["pdf2pic"];
+          ? ["pdf2pic","pdf-lib"]
+          : [];
         return deps.some(dep => id === dep || id.startsWith(dep + "/"));
       }
     }