+++ libs/skill-qrcode/rollup.config.cjs
@@ -21,7 +21,7 @@ module.exports = withNxDigipair(['esm', 'cjs'], config =>
     {      
       external: (id) => {
         const deps = config.format === 'cjs'
-          ? ["canvas"]
+          ? ["canvas", "qrcode", "jsqr"]
           : ["canvas"];
         return deps.some(dep => id === dep || id.startsWith(dep + "/"));
       }