+++ libs/skill-tensorflow/src/index.d.ts
@@ -1,1 +1,0 @@
-export * from './libs/skill-tensorflow/src/index';