+++ libs/skill-tensorflow/src/lib/skill-tensorflow.spec.ts
@@ -1,7 +1,0 @@
-import { skillTensorflow } from './skill-tensorflow';
-
-describe('skillTensorflow', () => {
-  it('should work', () => {
-    expect(skillTensorflow()).toEqual('skill-tensorflow');
-  });
-});